1) Users auth is partly implemented 2) JSON parsing errors are now visible 3) websockets test page

+++ src/main/java/com/devicehive/websockets/messagebus/local/subscriptions/CommandsSubscriptionManager.java
@@ -104,7 +104,7 @@ public class CommandsSubscriptionManager implements Serializable {
     }
 
 
-    public Session getClientSession(Integer commandId) {
+    public Session getClientSession(Long commandId) {
         return commandToClientSessionMap.get(commandId);
     }
